4[refactor] Contact 패키지 > ContacComment-> Comment  변경
