+++ src/main/java/com/team/final8teamproject/contact/Comment/dto/CommentRequest.java
@@ -1,7 +1,7 @@
-package com.team.final8teamproject.contact.contactComment.dto;
+package com.team.final8teamproject.contact.Comment.dto;
 
 
-import com.team.final8teamproject.contact.contactComment.entity.Comment;
+import com.team.final8teamproject.contact.Comment.entity.Comment;
 import com.team.final8teamproject.contact.entity.Inquiry;
 import lombok.AccessLevel;
 import lombok.Builder;
