+++ src/main/java/com/team/final8teamproject/contact/Comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
-package com.team.final8teamproject.contact.contactComment.repository;
+package com.team.final8teamproject.contact.Comment.repository;
 
-import com.team.final8teamproject.contact.contactComment.entity.Comment;
+import com.team.final8teamproject.contact.Comment.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment,Long>{
