+++ src/main/java/com/team/final8teamproject/contact/Comment/servive/CommentServiceImpl.java
@@ -1,8 +1,8 @@
-package com.team.final8teamproject.contact.contactComment.servive;
+package com.team.final8teamproject.contact.Comment.servive;
 
-import com.team.final8teamproject.contact.contactComment.dto.CommentRequest;
-import com.team.final8teamproject.contact.contactComment.entity.Comment;
-import com.team.final8teamproject.contact.contactComment.repository.CommentRepository;
+import com.team.final8teamproject.contact.Comment.dto.CommentRequest;
+import com.team.final8teamproject.contact.Comment.entity.Comment;
+import com.team.final8teamproject.contact.Comment.repository.CommentRepository;
 import com.team.final8teamproject.contact.entity.Inquiry;
 import com.team.final8teamproject.contact.service.InquiryServiceImpl;
